move InputOverlayKey class to separate file
